refactor: Add new category actions and refactor CategoryBadge to use constants
